requirements.txt
snowflake-connector-python
requests
datetime